requirements.txt
langchain-core 
langgraph>0.2.27
openai-whisper
groq
langchain-groq
langgraph-checkpoint-sqlite
sentence-transformers
jiwer
gradio